--- FoodBowl/Screens/Tab/Find/View/FindStoreViewController.swift
@@ -1,64 +1,0 @@
-//
-//  FindStoreViewController.swift
-//  FoodBowl
-//
-//  Created by COBY_PRO on 2023/01/18.
-//
-
-import UIKit
-
-import SnapKit
-import Then
-
-final class FindStoreViewController: BaseViewController {
-    var delegate: FindStoreViewControllerDelegate?
-
-    private lazy var storeResultTableView = UITableView().then {
-        $0.register(StoreInfoTableViewCell.self, forCellReuseIdentifier: StoreInfoTableViewCell.className)
-        $0.delegate = self
-        $0.dataSource = self
-        $0.separatorInset = UIEdgeInsets(top: 0, left: 0, bottom: 20, right: 0)
-        $0.backgroundColor = .mainBackground
-    }
-
-    override func setupLayout() {
-        view.addSubviews(storeResultTableView)
-
-        storeResultTableView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-        }
-    }
-}
-
-extension FindStoreViewController: UITableViewDataSource, UITableViewDelegate {
-    func tableView(_: UITableView, numberOfRowsInSection _: Int) -> Int {
-        return 5
-    }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView
-            .dequeueReusableCell(withIdentifier: StoreInfoTableViewCell.className, for: indexPath) as? StoreInfoTableViewCell
-        else { return UITableViewCell() }
-
-        cell.selectionStyle = .none
-        cell.storeNameLabel.text = "가게이름"
-        cell.storeFeedLabel.text = "100명이 후기를 남겼습니다."
-        cell.storeDistanceLabel.text = "10km"
-
-        return cell
-    }
-
-    func tableView(_: UITableView, heightForRowAt _: IndexPath) -> CGFloat {
-        return 64
-    }
-
-    func tableView(_: UITableView, didSelectRowAt _: IndexPath) {
-        let storeDetailViewController = StoreDetailViewController()
-        storeDetailViewController.title = "틈새라면"
-        delegate?.setStore(storeDetailViewController: storeDetailViewController)
-    }
-}
-
-protocol FindStoreViewControllerDelegate: AnyObject {
-    func setStore(storeDetailViewController: StoreDetailViewController)
-}
